pyproject.toml
[tool.poetry]
name = "assetmanagementservice"
version = "0.1.0"
description = ""
authors = ["Mio Lau <liurusi.101@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
#fastapi = "^0.75.1"
sanic = {extras = ["ext"], version = "^22.3.1"}
#hypercorn = "^0.13.2"
databases = {extras = ["aiomysql"], version = "^0.5.5"}
aiomysql = "^0.1.0"
dynaconf = "^3.1.8"
arrow = "^1.2.2"
ecdsa = "^0.17.0"
pycryptodome = "^3.14.1"
base58 = "^2.1.1"
stellar-sdk = "^8.0.0"
schema = "^0.7.5"
loguru = "^0.6.0"
redis = "^4.3.1"
types-redis = "^4.2.5"
sanic-scheduler = "^1.0.7"
telethon = {extras = ["socks"], version = "^1.24.0"}
python-socks = {extras = ["asyncio"], version = "^2.0.3"}

[tool.poetry.dev-dependencies]
#locust = "^2.8.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


#[[tool.poetry.source]]
#name = "aliyun"
#url = "https://mirrors.aliyun.com/pypi/simple/"
#default = true


[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple"
default = true